Consolidate report helpers
