+++ games/horizonzdr/hzdr.py
@@ -15,8 +15,7 @@ sys.path.insert(1, PARENT_DIRECTORY)
 from harness_utils.artifacts import ArtifactManager, ArtifactType
 from harness_utils.ocr_service import find_word
 from harness_utils.file_cleanup import remove_files
-from harness_utils.resolution_formatting import format_resolution
-from harness_utils.time_conversion import seconds_to_milliseconds
+from harness_utils.report import format_resolution, seconds_to_milliseconds
 from harness_utils.output_logging import setup_logging
 from harness_utils.report_writing import write_report_json
 from harness_utils.process import terminate_process
