+++ games/tinytinaswonderland/tinytinaswonderland.py
@@ -17,8 +17,7 @@ sys.path.insert(1, PARENT_DIRECTORY)
 
 from harness_utils.artifacts import ArtifactManager, ArtifactType
 from harness_utils.ocr_service import find_word
-from harness_utils.resolution_formatting import format_resolution
-from harness_utils.time_conversion import seconds_to_milliseconds
+from harness_utils.report import format_resolution, seconds_to_milliseconds
 from harness_utils.report_writing import write_report_json
 from harness_utils.process import terminate_process
 from harness_utils.steam import exec_steam_game, get_build_id
