--- harness_utils/resolution_formatting.py
@@ -1,6 +1,0 @@
-"""Resolution formatting helpers for harnesses."""
-
-
-def format_resolution(width: int, height: int) -> str:
-    """Return a resolution in WxH format."""
-    return f"{width}x{height}"
