--- harness_utils/time_conversion.py
@@ -1,6 +1,0 @@
-"""Time unit conversion helpers for harnesses."""
-
-
-def seconds_to_milliseconds(seconds: float | int) -> int:
-    """Convert seconds to milliseconds."""
-    return round(seconds * 1000)
