+++ non_games/cinebench_2024/cinebench_2024.py
@@ -13,9 +13,8 @@ from cinebench_2024_utils import friendly_test_name, get_score
 PARENT_DIRECTORY = str(Path(__file__).resolve().parent.parent.parent)
 sys.path.insert(1, PARENT_DIRECTORY)
 
-from harness_utils.time_conversion import seconds_to_milliseconds
+from harness_utils.report import seconds_to_milliseconds, write_report_json
 from harness_utils.output_logging import setup_logging
-from harness_utils.report_writing import write_report_json
 
 CINEBENCH_PATH = r"C:\Cinebench2024\Cinebench.exe"
 GPU_TEST = "g_CinebenchGpuTest=true"
