+++ non_games/dawntrail/Dawntrail.py
@@ -15,9 +15,8 @@ from harness_utils.paths import network_drive_path
 from harness_utils.artifacts import ArtifactManager, ArtifactType
 from harness_utils.input import user
 from harness_utils.ocr_service import find_word
-from harness_utils.time_conversion import seconds_to_milliseconds
+from harness_utils.report import seconds_to_milliseconds, write_report_json
 from harness_utils.output_logging import setup_logging
-from harness_utils.report_writing import write_report_json
 from harness_utils.process import terminate_process
 
 SCRIPT_DIRECTORY = Path(__file__).resolve().parent
