+++ non_games/godot_compile/godot_compile.py
@@ -20,7 +20,7 @@ sys.path.insert(1, PARENT_DIRECTORY)
 from godot_compile_utils import current_time_ms
 
 from harness_utils.output_logging import setup_logging
-from harness_utils.report_writing import write_report_json
+from harness_utils.report import write_report_json
 
 SCRIPT_DIRECTORY = Path(__file__).resolve().parent
 LOG_DIRECTORY = SCRIPT_DIRECTORY / "run"
