+++ non_games/gravitymark/gravitymark.py
@@ -17,7 +17,7 @@ PARENT_DIRECTORY = str(Path(__file__).resolve().parent.parent.parent)
 sys.path.insert(1, PARENT_DIRECTORY)
 
 from harness_utils.output_logging import setup_logging
-from harness_utils.report_writing import write_report_json
+from harness_utils.report import write_report_json
 
 GRAVITYMARK_PATH = Path("C:/", "Program Files", "GravityMark", "bin")
 GRAVITYMARK_EXE = GRAVITYMARK_PATH / "GravityMark.exe"
