+++ non_games/procyon/procyon_ai_img_gen/ulprocai_img_gen.py
@@ -20,9 +20,8 @@ PARENT_DIRECTORY = str(Path(__file__).resolve().parent.parent.parent.parent)
 sys.path.insert(1, PARENT_DIRECTORY)
 
 from harness_utils.artifacts import ArtifactManager, ArtifactType
-from harness_utils.time_conversion import seconds_to_milliseconds
+from harness_utils.report import seconds_to_milliseconds, write_report_json
 from harness_utils.output_logging import setup_logging
-from harness_utils.report_writing import write_report_json
 from non_games.procyon.procyoncmd import (
     get_cuda_devices,
     get_openvino_devices,
